+++ docs/history.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - The `env show`, `dep hash`, and all `dep show` commands now honor context formatting
 - Decreasing verbosity now has no affect on output that should always be displayed
 - Handle more edge cases in the `setuptools` migration script
+- Environments now respect user defined environment variables for context formatting
 - Allow `extra-dependencies` in environment overrides
 - Depend on `packaging` explicitly rather than relying on it being a transitive dependency of Hatchling